refactor: simplified mappo system design imports
