+++ mava/components/jax/building/__init__.py
@@ -14,3 +14,18 @@
 # limitations under the License.
 
 """Builder components for Mava systems."""
+
+from mava.components.jax.building.adders import (
+    ParallelSequenceAdder,
+    ParallelSequenceAdderSignature,
+)
+from mava.components.jax.building.data_server import OnPolicyDataServer
+from mava.components.jax.building.datasets import TrajectoryDataset
+from mava.components.jax.building.distributor import Distributor
+from mava.components.jax.building.environments import ParallelExecutorEnvironmentLoop
+from mava.components.jax.building.loggers import Logger
+from mava.components.jax.building.networks import DefaultNetworks
+from mava.components.jax.building.parameter_client import (
+    ExecutorParameterClient,
+    TrainerParameterClient,
+)
