+++ mava/components/jax/executing/action_selection.py
@@ -1,0 +1,75 @@
+# python3
+# Copyright 2021 InstaDeep Ltd. All rights reserved.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+"""Execution components for system builders"""
+
+from dataclasses import dataclass
+
+import jax
+
+from mava.components.jax import Component
+from mava.core_jax import SystemExecutor
+
+
+@dataclass
+class ExecutorSelectActionProcessConfig:
+    pass
+
+
+class FeedforwardExecutorSelectAction(Component):
+    def __init__(
+        self,
+        config: ExecutorSelectActionProcessConfig = ExecutorSelectActionProcessConfig(),
+    ):
+        """_summary_
+
+        Args:
+            config : _description_.
+        """
+        self.config = config
+
+    # Select actions
+    def on_execution_select_actions(self, executor: SystemExecutor) -> None:
+        """Summary"""
+        executor.store.actions_info = {}
+        executor.store.policies_info = {}
+        for agent, observation in executor.store.observations.items():
+            action_info, policy_info = executor.select_action(agent, observation)
+            executor.store.actions_info[agent] = action_info
+            executor.store.policies_info[agent] = policy_info
+
+    # Select action
+    def on_execution_select_action_compute(self, executor: SystemExecutor) -> None:
+        """Summary"""
+
+        agent = executor.store.agent
+        network = executor.store.networks["networks"][
+            executor.store.agent_net_keys[agent]
+        ]
+
+        observation = executor.store.observation.observation.reshape((1, -1))
+        rng_key, executor.store.key = jax.random.split(executor.store.key)
+
+        # TODO (dries): We are currently using jit in the networks per agent.
+        # We can also try jit over all the agents in a for loop. This would
+        # allow the jit function to save us even more time.
+        executor.store.action_info, executor.store.policy_info = network.get_action(
+            observation, rng_key
+        )
+
+    @property
+    def name(self) -> str:
+        """_summary_"""
+        return "action_selector"
