+++ mava/components/jax/executing/base.py
@@ -16,9 +16,7 @@
 """Execution components for system builders"""
 
 from dataclasses import dataclass
-from typing import Any, Dict, List, Optional, Union
-
-import jax
+from typing import List, Optional, Union
 
 from mava.components.jax import Component
 from mava.core_jax import SystemBuilder, SystemExecutor
@@ -27,7 +25,7 @@ from mava.utils.sort_utils import sample_new_agent_keys, sort_str_num
 
 
 @dataclass
-class ExecutorProcessConfig:
+class ExecutorInitConfig:
     network_sampling_setup: Union[
         List, enums.NetworkSampler
     ] = enums.NetworkSampler.fixed_agent_networks
@@ -35,7 +33,7 @@ class ExecutorProcessConfig:
 
 
 class ExecutorInit(Component):
-    def __init__(self, config: ExecutorProcessConfig = ExecutorProcessConfig()):
+    def __init__(self, config: ExecutorInitConfig = ExecutorInitConfig()):
         """_summary_
 
         Args:
@@ -131,138 +129,3 @@ class ExecutorInit(Component):
     def name(self) -> str:
         """_summary_"""
         return "executor_init"
-
-
-@dataclass
-class ExecutorObserveProcessConfig:
-    pass
-
-
-class FeedforwardExecutorObserve(Component):
-    def __init__(
-        self, config: ExecutorObserveProcessConfig = ExecutorObserveProcessConfig()
-    ):
-        """_summary_
-
-        Args:
-            config : _description_.
-        """
-        self.config = config
-
-    # Observe first
-    def on_execution_observe_first(self, executor: SystemExecutor) -> None:
-        """_summary_
-
-        Args:
-            executor : _description_
-        """
-        if not executor.store.adder:
-            return
-
-        "Select new networks from the sampler at the start of each episode."
-        agents = sort_str_num(list(executor.store.agent_net_keys.keys()))
-        (
-            executor.store.network_int_keys_extras,
-            executor.store.agent_net_keys,
-        ) = sample_new_agent_keys(
-            agents,
-            executor.store.network_sampling_setup,
-            executor.store.net_keys_to_ids,
-        )
-        executor.store.extras[
-            "network_int_keys"
-        ] = executor.store.network_int_keys_extras
-
-        executor.store.adder.add_first(executor.store.timestep, executor.store.extras)
-
-    # Observe
-    def on_execution_observe(self, executor: SystemExecutor) -> None:
-        """_summary_
-
-        Args:
-            executor : _description_
-        """
-        if not executor.store.adder:
-            return
-
-        actions_info = executor.store.actions_info
-        policies_info = executor.store.policies_info
-
-        adder_actions: Dict[str, Any] = {}
-        executor.store.next_extras["policy_info"] = {}
-        for agent in actions_info.keys():
-            adder_actions[agent] = {
-                "actions_info": actions_info[agent],
-            }
-            executor.store.next_extras["policy_info"][agent] = policies_info[agent]
-
-        executor.store.next_extras[
-            "network_int_keys"
-        ] = executor.store.network_int_keys_extras
-
-        executor.store.adder.add(
-            adder_actions, executor.store.next_timestep, executor.store.next_extras
-        )
-
-    # Update the executor variables.
-    def on_execution_update(self, executor: SystemExecutor) -> None:
-        """Update the policy variables."""
-        if executor.store.executor_parameter_client:
-            executor.store.executor_parameter_client.get_async()
-
-    @property
-    def name(self) -> str:
-        """_summary_"""
-        return "executor_observe"
-
-
-@dataclass
-class ExecutorSelectActionProcessConfig:
-    pass
-
-
-class FeedforwardExecutorSelectAction(Component):
-    def __init__(
-        self,
-        config: ExecutorSelectActionProcessConfig = ExecutorSelectActionProcessConfig(),
-    ):
-        """_summary_
-
-        Args:
-            config : _description_.
-        """
-        self.config = config
-
-    # Select actions
-    def on_execution_select_actions(self, executor: SystemExecutor) -> None:
-        """Summary"""
-        executor.store.actions_info = {}
-        executor.store.policies_info = {}
-        for agent, observation in executor.store.observations.items():
-            action_info, policy_info = executor.select_action(agent, observation)
-            executor.store.actions_info[agent] = action_info
-            executor.store.policies_info[agent] = policy_info
-
-    # Select action
-    def on_execution_select_action_compute(self, executor: SystemExecutor) -> None:
-        """Summary"""
-
-        agent = executor.store.agent
-        network = executor.store.networks["networks"][
-            executor.store.agent_net_keys[agent]
-        ]
-
-        observation = executor.store.observation.observation.reshape((1, -1))
-        rng_key, executor.store.key = jax.random.split(executor.store.key)
-
-        # TODO (dries): We are currently using jit in the networks per agent.
-        # We can also try jit over all the agents in a for loop. This would
-        # allow the jit function to save us even more time.
-        executor.store.action_info, executor.store.policy_info = network.get_action(
-            observation, rng_key
-        )
-
-    @property
-    def name(self) -> str:
-        """_summary_"""
-        return "executor"
